Complete rewrite of rules logic.
Previously every move was checked against lists of pawns and fences to check legality.
Now, the board is represented with a graph. Placing fences disconnects edges in the graph.
Checking for legal moves is simpler and faster.

Restructure code to modularise it.

Added command line option for randomness factor (introduces errors in the Machine player)